backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
openai==1.50.2
requests==2.32.3